JWT周りのリファクタリング
