+++ src/domain/user.go
@@ -2,16 +2,20 @@ package domain
 
 import (
 	"errors"
+	"time"
+
+	"github.com/dgrijalva/jwt-go"
 )
 
 type User struct {
-	ID       int
-	Email    string
-	Password string
+	ID         int
+	Email      string
+	Password   string
+	ValidEmail bool
 }
 
-func NewUser(id int, email string, password string) (*User, error) {
-	u := &User{ID: id, Email: email, Password: password}
+func NewUser(email string, password string) (*User, error) {
+	u := &User{Email: email, Password: password}
 	err := u.validation()
 	if err != nil {
 		return nil, err
@@ -20,9 +24,6 @@ func NewUser(id int, email string, password string) (*User, error) {
 }
 
 func (u *User) validation() error {
-	if u.ID == 0 {
-		return errors.New("IDを格納してください。")
-	}
 	if u.Email == "" {
 		return errors.New("Emailを格納してください。")
 	}
@@ -30,4 +31,19 @@ func (u *User) validation() error {
 		return errors.New("Passwordを格納してください。")
 	}
 	return nil
+}
+
+func (u *User) CreateJWT(id int) (string, error) {
+	token := jwt.NewWithClaims(jwt.GetSigningMethod("HS256"), jwt.MapClaims{
+		"id":    id,
+		"email": u.Email,
+		"exp":   time.Now().Add(time.Hour * 24).Unix(),
+		"iat":   time.Now(),
+	})
+	tokenstring, err := token.SignedString([]byte("foobar"))
+
+	if err != nil {
+		return "", err
+	}
+	return tokenstring, nil
 }
