+++ src/usecase/user_test.go
@@ -3,7 +3,6 @@ package usecase
 import (
 	"testing"
 
-	"github.com/dgrijalva/jwt-go"
 	. "github.com/smartystreets/goconvey/convey"
 )
 
@@ -17,29 +16,6 @@ func TestAdd(t *testing.T) {
 	})
 }
 
-func TestFetchJWT(t *testing.T) {
-	Convey("FetchJWT()でtokenが返却されることを確認する", t, func() {
-		r := new(testRepo)
-		u := &UserUsecase{r}
-		t, _ := u.FetchJWT("ex@example.com", "password")
-		token, _ := jwt.Parse(t, func(token *jwt.Token) (interface{}, error) {
-			return []byte("foobar"), nil
-		})
-
-		Convey("emailが格納されていること", func() {
-			So(token.Claims.(jwt.MapClaims)["email"], ShouldEqual, "ex@example.com")
-		})
-		Convey("idが格納されていること", func() {
-			So(token.Claims.(jwt.MapClaims)["id"], ShouldEqual, 1)
-		})
-		Convey("expが格納されていること", func() {
-			So(token.Claims.(jwt.MapClaims)["exp"], ShouldNotBeNil)
-		})
-		Convey("iatが格納されていること", func() {
-			So(token.Claims.(jwt.MapClaims)["iat"], ShouldNotBeNil)
-		})
-	})
-}
 func TestCreateHashPassword(t *testing.T) {
 	Convey("hash化されているか検証する", t, func() {
 		hash := createHashPassword("password")
